fix(userdata): clean up docker cruft weekly
Docker is fairly messy when you pull a new version of an image from a
repository, or build a new one for that matter. It accidentally creates
a lot of orphaned containers and images. This is a port of some
utilities from my dotfiles that should help with disk usage on the
cluster machines. This does not currently clean the builder.